+++ substrafl/algorithms/pytorch/torch_base_algo.py
@@ -390,12 +390,14 @@ class TorchAlgo(Algo):
             {
                 "model": str(type(self._model)),
                 "criterion": str(type(self._criterion)),
-                "optimizer": None
-                if self._optimizer is None
-                else {
-                    "type": str(type(self._optimizer)),
-                    "parameters": self._optimizer.defaults,
-                },
+                "optimizer": (
+                    None
+                    if self._optimizer is None
+                    else {
+                        "type": str(type(self._optimizer)),
+                        "parameters": self._optimizer.defaults,
+                    }
+                ),
                 "scheduler": None if self._scheduler is None else str(type(self._scheduler)),
             }
         )