+++ substrafl/nodes/aggregation_node.py
@@ -1,5 +1,6 @@
 import uuid
 from typing import Dict
+from typing import List
 from typing import Optional
 from typing import Set
 from typing import TypeVar
@@ -7,11 +8,11 @@ from typing import TypeVar
 import substra
 
 from substrafl.dependency import Dependency
-from substrafl.nodes.node import InputIdentifiers
-from substrafl.nodes.node import Node
-from substrafl.nodes.node import OperationKey
-from substrafl.nodes.node import OutputIdentifiers
+from substrafl.nodes.protocol import AggregationNodeProtocol
 from substrafl.nodes.references.shared_state import SharedStateRef
+from substrafl.nodes.schemas import InputIdentifiers
+from substrafl.nodes.schemas import OperationKey
+from substrafl.nodes.schemas import OutputIdentifiers
 from substrafl.remote.operations import RemoteOperation
 from substrafl.remote.register import register_function
 from substrafl.remote.remote_struct import RemoteStruct
@@ -20,11 +21,15 @@ from substrafl.schemas import TaskType
 SharedState = TypeVar("SharedState")
 
 
-class AggregationNode(Node):
+class AggregationNode(AggregationNodeProtocol):
     """The node which applies operations to the shared states which are received from ``TrainDataNode``
     data operations.
     The result is sent to the ``TrainDataNode`` and/or ``TestDataNode`` data operations.
     """
+
+    def __init__(self, organization_id: str):
+        self.organization_id = organization_id
+        self.tasks: List[Dict] = []
 
     def update_states(
         self,
@@ -161,3 +166,13 @@ class AggregationNode(Node):
                 task["function_key"] = function_key
 
         return cache
+
+    def summary(self) -> dict:
+        """Summary of the class to be exposed in the experiment summary file
+
+        Returns:
+            dict: a json-serializable dict with the attributes the user wants to store
+        """
+        return {
+            "organization_id": self.organization_id,
+        }