+++ substrafl/strategies/fed_avg.py
@@ -8,11 +8,11 @@ import numpy as np
 
 from substrafl.algorithms.algo import Algo
 from substrafl.exceptions import EmptySharedStatesError
-from substrafl.nodes.aggregation_node import AggregationNode
+from substrafl.nodes import AggregationNodeProtocol
+from substrafl.nodes import TestDataNodeProtocol
+from substrafl.nodes import TrainDataNodeProtocol
 from substrafl.nodes.references.local_state import LocalStateRef
 from substrafl.nodes.references.shared_state import SharedStateRef
-from substrafl.nodes.test_data_node import TestDataNode
-from substrafl.nodes.train_data_node import TrainDataNode
 from substrafl.remote import remote
 from substrafl.strategies.schemas import FedAvgAveragedState
 from substrafl.strategies.schemas import FedAvgSharedState
@@ -28,8 +28,8 @@ class FedAvg(Strategy):
     passes on each client, aggregating updates by computing their means and
     distributing the consensus update to all clients. In FedAvg, strategy is
     performed in a centralized way, where a single server or
-    ``AggregationNode`` communicates with a number of clients ``TrainDataNode``
-    and ``TestDataNode``.
+    ``AggregationNodeProtocol`` communicates with a number of clients ``TrainDataNodeProtocol``
+    and ``TestDataNodeProtocol``.
 
     Formally, if :math:`w_t` denotes the parameters of the model at round
     :math:`t`, a single round consists in the following steps:
@@ -79,8 +79,8 @@ class FedAvg(Strategy):
     def perform_round(
         self,
         *,
-        train_data_nodes: List[TrainDataNode],
-        aggregation_node: AggregationNode,
+        train_data_nodes: List[TrainDataNodeProtocol],
+        aggregation_node: AggregationNodeProtocol,
         round_idx: int,
         clean_models: bool,
         additional_orgs_permissions: Optional[set] = None,
@@ -93,9 +93,9 @@ class FedAvg(Strategy):
             - perform a local update (train on n mini-batches) of the models on each train data nodes
 
         Args:
-            train_data_nodes (typing.List[TrainDataNode]): List of the nodes on which to perform
+            train_data_nodes (typing.List[TrainDataNodeProtocol]): List of the nodes on which to perform
                 local updates.
-            aggregation_node (AggregationNode): Node without data, used to perform
+            aggregation_node (AggregationNodeProtocol): Node without data, used to perform
                 operations on the shared states of the models
             round_idx (int): Round number, it starts at 0.
             clean_models (bool): Clean the intermediary models of this round on the Substra platform.
@@ -138,15 +138,15 @@ class FedAvg(Strategy):
 
     def perform_evaluation(
         self,
-        test_data_nodes: List[TestDataNode],
-        train_data_nodes: List[TrainDataNode],
+        test_data_nodes: List[TestDataNodeProtocol],
+        train_data_nodes: List[TrainDataNodeProtocol],
         round_idx: int,
     ):
         """Perform evaluation on test_data_nodes.
 
         Args:
-            test_data_nodes (List[TestDataNode]): test data nodes to perform the prediction from the algo on.
-            train_data_nodes (List[TrainDataNode]): train data nodes the model has been trained
+            test_data_nodes (List[TestDataNodeProtocol]): test data nodes to perform the prediction from the algo on.
+            train_data_nodes (List[TrainDataNodeProtocol]): train data nodes the model has been trained
                 on.
             round_idx (int): round index.
         """
@@ -167,7 +167,7 @@ class FedAvg(Strategy):
             test_data_node.update_states(
                 traintask_id=local_state.key,
                 operation=self.evaluate(
-                    data_samples=test_data_node.test_data_sample_keys,
+                    data_samples=test_data_node.data_sample_keys,
                     _algo_name=f"Evaluating with {self.__class__.__name__}",
                 ),
                 round_idx=round_idx,
@@ -225,7 +225,7 @@ class FedAvg(Strategy):
 
     def _perform_local_updates(
         self,
-        train_data_nodes: List[TrainDataNode],
+        train_data_nodes: List[TrainDataNodeProtocol],
         current_aggregation: Optional[SharedStateRef],
         round_idx: int,
         aggregation_id: str,
@@ -236,7 +236,7 @@ class FedAvg(Strategy):
         on each train data nodes.
 
         Args:
-            train_data_nodes (typing.List[TrainDataNode]): List of the organizations on which to perform
+            train_data_nodes (typing.List[TrainDataNodeProtocol]): List of the organizations on which to perform
             local updates current_aggregation (SharedStateRef, Optional): Reference of an aggregation operation to
                 be passed as input to each local training
             round_idx (int): Round number, it starts at 1.